Rename to preflightChecks
