+++ packages/sdk/src/lib/dataProtectorSharing/removeCollection.ts
@@ -10,7 +10,7 @@ import {
   SuccessWithTransactionHash,
 } from '../types/index.js';
 import { getSharingContract } from './smartContract/getSharingContract.js';
-import { onlyCollectionOperator } from './smartContract/preFlightCheck.js';
+import { onlyCollectionOperator } from './smartContract/preflightChecks.js';
 
 export const removeCollection = async ({
   iexec = throwIfMissing(),
