+++ packages/sdk/src/lib/dataProtectorSharing/smartContract/preflightChecks.ts
